+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Unreleased
 
+- Add one-click install option for ocaml-lsp-server (#1725)
+
 ## 1.28.0
 
 - Add `.re` file extension support. (#1685)